close unused file descriptors
Without this, a cgi reading from stdin will block because it still has the
write end of the pipe open. This matters to cgi scripts using POST requests
with multipart/form-data, as the examples in pages/index.cgi and
pages/getfile.cgi (let me know if you want me to remove these two files from
the commit, they are just examples to demonstrate the problem).